Bump tinyMCE default version number to 8.0.1
This is a major release upgrade, with a couple breaking changes (see https://www.tiny.cloud/docs/tinymce/latest/migration-from-7x/).

As language pack for version 8 do not exist yet, we currently keep using those from 7 for any version >= 7. This should go away as soon as an own language pack is released for 8.